oppretter domainmappe med typer og legger til mock paa regioner
