+++ .gitignore
@@ -12,6 +12,7 @@
 .idea/*
 # misc
 .DS_Store
+.env
 .env.local
 .env.development.local
 .env.test.local
